feat: add an optional handler parameter to EventNotifier.prototype.dispatch method.
EventNotifier.prototype.dispatch now supports an optional third parameter onFinished.
The parameter is a function that is called after an event dispatch and reports all values returned by the event handlers of the raised event or an error thrown by one of the event handlers.
If you want to explicitly handle errors thrown by the event handlers, you should pass a proper handler function as the third parameter.